refactor: move pure utility functions out of UI components (#1194)
