+++ surfsense_web/lib/comments/utils.ts
@@ -1,0 +1,4 @@
+export function convertRenderedToDisplay(contentRendered: string): string {
+	// Convert @{DisplayName} format to @DisplayName for editing
+	return contentRendered.replace(/@\{([^}]+)\}/g, "@$1");
+}
