+++ surfsense_web/lib/layout-events.ts
@@ -1,0 +1,1 @@
+export const SLIDEOUT_PANEL_OPENED_EVENT = "slideout-panel-opened";
